+++ app/base/config.js
@@ -239,7 +239,7 @@ angular.module('cmConfig',[])
 
             appLinks: JSON.parse('<%= appLinks %>'),
             appProtocol: '<%= appProtocol %>',
-
+            appIcon: '<%= appIcon %>',
 
             autoLoginData: {
                 'Dumpuser local': {