+++ core/comps/contacts/drtv-contact-import.js
@@ -11,16 +11,10 @@
 
 angular.module('cmContacts')
 .directive('cmContactImport', [
-
-    'cmContactsModel', 'cmUtil', 'cmModal', 'cmNotify',
-    'cmLocalContacts', 'cmConversationFactory', 'cmIdentityFactory', 'cmTranslate',
-    'cmUserModel', 'cmLoader',
-    '$rootScope', '$q', '$filter',
-
-    function(cmContactsModel, cmUtil, cmModal, cmNotify,
-             cmLocalContacts, cmConversationFactory, cmIdentityFactory, cmTranslate,
-             cmUserModel, cmLoader,
-             $rootScope, $q, $filter){
+    'cmContactsModel', 'cmUtil', 'cmNotify', 'cmLocalContacts', 'cmLoader', 'cmModalContactImport',
+    '$rootScope', '$q',
+    function(cmContactsModel, cmUtil, cmNotify, cmLocalContacts, cmLoader, cmModalContactImport,
+             $rootScope, $q){
 
         return {
             restrict:       'AE',
@@ -151,8 +145,6 @@ angular.module('cmContacts')
 
                     $scope.validateForm().then(
                         function(objectChange) {
-
-                            // everything is fine let's add the contact
                             cmContactsModel
                             .addContact({
                                 identity: objectChange
@@ -160,45 +152,14 @@ angular.module('cmContacts')
                             .then(
                                 function (data) {
                                     loader.stop();
-                                    var identity = cmIdentityFactory.create(data.identity, true);
-
-                                    var messageData = {
-                                        from: cmUserModel.data.identity.getDisplayName(),
-                                        to: identity.getDisplayName()
-                                    };
-
-                                    return cmModal.confirm({
-                                                title: '',
-                                                text:  'CONTACT.IMPORT.NOTIFICATION.CONFIRMATION',
-                                                html:  '<textarea cm-resize-textarea class="confirm-textarea" ng-model="data.message"> </textarea>',
-                                                data:  {
-                                                    message: $filter('cmTranslate')('CONTACT.IMPORT.NOTIFICATION.MESSAGE',messageData)
-                                                }
-                                            })
+                                    return new cmModalContactImport(data);
                                 },
                                 function () {
                                     loader.stop();
                                     cmNotify.error('CONTACT.INFO.ERROR.SAVE', {ttl: 5000});
-                                    return $q.reject()
+                                    return $q.reject();
                                 }
                             )
-                            .then(function(modal_scope){
-                                var conversation =  cmConversationFactory
-                                                    .create()
-                                                    .addRecipient(identity)
-                                                    .disableEncryption();
-
-                                return  conversation
-                                        .save()
-                                        .then(function(){
-                                            return  conversation
-                                                    .messages
-                                                    .create({conversation:conversation,text:modal_scope.data.message})
-                                                    .setPublicData(['text'])
-                                                    .save()
-                                        })
-
-                            })
                             .finally(function(){
                                 loader.stop();
                                 $scope.gotoContactList();