+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-#0.4.2
+# BS 21.11.2014
+* fix edit icon in contact edit view, disable if contact is cameo member
+* fix edit own identity CAM-1340
+
+# BS 20.11.2014
+* add BrowserNotificationService
+* fix invite service after importing a local contact
+* fix pending contacts in contacts-list CAM-1378
+
+0.4.2.
 * framework major update
 ** angularjs 1.2.15 upgraded to 1.3.2
 ** removed fastclick (ngTouch handle the 300ms click delay)
@@ -15,6 +24,7 @@
 ** phone and email validation now accors after 3 secs. (before 1 secs.)
 * phonegap
 ** new exit modal for leaving app on back button
+** device menubutton toggle the menu in app
 * performance
 ** webworker for verify now cache the results
 ** through major framework update route changes raised up 50%